Video thumbnails
When a video thumbnail is requested, an http server listens on an
ephemeral random port and serves only one ref to the thumbnailing
command (for now ffmpeg only). The ffmpeg process read the file from
the http server and outputs the thumbnail image without resizing on
its stdout which is read back by camlistore and cached. The file then
goes through the original image processing/resizing process.

Src is now provided in describe blob request.
To differentiate from image the parameter v=true is added to the
original thumbnail url. However, the size of the image is not known
at this time. In a later version we could extract information about
media files in the indexer (size/resolution, presence of audio and/or
video streams, format, codecs,...) with ffprobe.

Questions :
server/image.go : should the cache hit/miss of the original image
thumbnail before resizing be counted in thumbCache{Hit,Miss} ?

Change-Id: Ic960222d55e3eedb55b076015571fd33595548ae